Buffs the LP hardsuits (#3263)
- These are their primary gear, they better be good in comparison to
others.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This buffs the armor values of the LP hardsuits and adds slowdown
reduction and a reagent scanner to the medical hardsuit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The Ranger's whole shtick is that it revolves around the faux ert team.
they have little in the way of guns so their hardsuits are important.
They were meant to stay in level with other similar hardsuits but have
remained unmaintained.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:phoaly
balance: Buffed LP hardsuits
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->