Hardliner Rework DLC: Adds the Hardliner Elite Suit (#3394)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Hi! Kept you waiting, huh?

Does what it says on the tin, adding this beauty to the Hardliner's
armory!


![image](https://github.com/user-attachments/assets/71c00b2f-4814-439d-97c8-555d6f88a86a)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
We needed an elite analogue for the splinter Gorlex factions, and this
is the first one to kick things off!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
add: A new and refurbished Elite Syndicate suit has just been unveiled!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->