Adds NGR stamps and secret docs. (#3550)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

![obraz](https://github.com/user-attachments/assets/1d16e84a-c00b-49ad-8a50-d469bc1d598f)

![obraz](https://github.com/user-attachments/assets/38fa6b7a-65ec-4f8a-a7c4-51a12059a1e8)

Adds that stuff to the Hyena.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

There is an NGR combat ship and a Hyena replacement in the works rn.
Good time to add some fluff.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: Added stamps for New Gorlex Republic.
add: Added secret documents for New Gorlex Republic.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->