Being thrown over lava now doesn't cause you to fall into it (#3577)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Buff to jump boots and kepori trying to make the beef jump

## Why It's Good For The Game

im. over the lava.i m oiver it. it cant touch me. because im. not in it.

## Changelog

:cl:
balance: being thrown over lava by tackling, jump boots, or with
assistance now fully clears the lava instead of burning you horribly and
setting you on fire
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->