Fixes searching with the orbit menu (#3629)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR modifies the search algorithm for points of interests so that
the real name of a character, e.g. "Kristina Dryden" is what gets passed
to the search function, instead of a character's "full name", e.g. "calm
human woman".


![vf7sJJ034D](https://github.com/user-attachments/assets/59b9381b-f49d-4564-a8b7-e33653706693)

The sorting algorithm seems to take jobs as the sorting parameter,
instead of real names. As of right now, the portion of code that causes
this to occur still eludes me.

## Why It's Good For The Game

This felt like a minor oversight to me. It feels clunky to search for a
character's name and have the search return no results because the
search function had been looking for a character's _attributes_ instead.

## Changelog

:cl:
fix: fixed the orbit menu search function so that names work again
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->