Qdel Log Hotfix (#3633)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
#3558 accidentally ported the JSON log support for qdel logs, despite
the fact we still use old normal logs. This moves us back to the version
we support, and also adds support

## Why It's Good For The Game
qdel logs are nice to have for gauging prevalence of harddels

## Changelog

:cl:
server: qdel logs work again
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->