Edits the name of plasma and plastitanium shards (#3636)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Changes the name of plasmaglass shards from "purple shard" to
"plasmaglass shard" and the name of plastitanium glass shards from
"beige shard" to "plastitanium glass shard".

## Why It's Good For The Game

It does not fit with the current naming convention for items for these
shards to be vaguely named based on their color. This change would help
these shards make more sense.

## Changelog

:cl:
add: Added better names for some material shards
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->