Console examine_more viewability change (#3553)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR makes it so that you can view a console, laptop, a wideband log,
the screen on an APC, and the gasses measured from an air alarm, from
one tile away. You can not interact with the UI unless you are right
next to it. To do this, the examine_more function is used.

## Why It's Good For The Game

It makes sense to be able to see screens without having to be right next
to them. You will still be unable to interact with them, but you no
longer need to be right next to them just to open up the UI. Laptops
have also always required holding them in your hands to open up the UI,
and this change should address that so that this is no longer necessary.

## Changelog

:cl:
add: Added ability to see screens using the examine_more function
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->