Unstations Robotic Hud Alerts (#3664)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR kills a few stragglers in alert descriptions from the Great
Unstation that reference areas of "the station" when speaking about low
batteries or lockdowns in cyborgs or mechs. Also replaces an Ethereal
reference with our other beloved lightbulbs.

## Why It's Good For The Game

Robotics? Dormitories? AI satellite? What are you talking about,
recruit? It's 505 FSC on the frontier, and we're about to fly head-first
into an electrical storm.

## Changelog

:cl:
tweak: unstations some of the hud alerts.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Geoengi <139566513+Geoengi@users.noreply.github.com>