Vox Caps and Hard Hats Resprite (#3754)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Resprites Hard Hats to be in line with the better hard hats made by
Imaginos and adapted to Vox by me.
Similarly resprites Soft Caps to be based off the NT softcaps, also made
by Imaginos. Done with permission.

Softcaps were not resprited for humans in this PR as a project by others
is already underway to do so, and I do not want to step on toes in the
process.


![image](https://github.com/user-attachments/assets/ab6951de-5118-4553-88fd-9a1d66aaafd1)


![image](https://github.com/user-attachments/assets/898126f5-7ee0-46d0-b7ed-ea3cc13eeb05)

Example hat adaptations.

## Why It's Good For The Game

Resprites older cruft sprites for Vox to be more in line with the
current artstyle while.

## Changelog

:cl:
add: New sprites for hard hats and soft caps for Vox! Flipped states and
on/off states included.
/:cl: