Another round of fixes (#3712)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Fixes a bunch of small, mostly unrelated runtimes, errors, and possibly
a harddel. Most non-issue runtimes found in the [latest consolidated
runtime
logs](https://cdn.shiptest.net/logs/2024/11/03-runtimes-consolidated.log.gz).

Also standardizes `examine_more()` because it was a bit of a mess as is
that led to runtimes.

Fixes: #3671
Fixes: #3654

We honestly have so much fewer runtimes than we used to, it's unreal. We
might be close to being in the double-digits worth of preventable
runtimes if we keep up

:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->