Succumb Code Touchup (#3711)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Touches up saycode related to succumbing just a little bit. Originally
was intended to actually solve a problem, but it didn't, and I figured
it was fine to add the tweaks I made anyways for the hell of it.

Also made the dramatic "last words" screen text a bit more consistent.
Yes, I know it was for drama reasons, but this is far better than having
to use an entire get_hearers call for it, and I doubt anyone will notice
a difference.

## Why It's Good For The Game
Succumb code was kind of a mess, so I cleaned it up. Honestly, I don't
think this should be in say() at all, but hey, I just work here.

## Changelog

:cl:
tweak: Whispering (as in, directly using the verb or typing `#` before
your message) will now skip the succumb confirmation prompt, while
normal talking will still continue to prompt if you want to succumb.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Sun-Soaked <45698967+Sun-Soaked@users.noreply.github.com>