Replaces the Shetland's TEG and combustion engines with a turbine (#4649)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Replaces the Shetland's thermo-electric generator and combustion engines
with a turbine.

<details>
<summary>Mapping</summary>

![image](https://github.com/user-attachments/assets/c9d712c0-4e1d-4321-94ad-3b8e480d344c)

![image](https://github.com/user-attachments/assets/cb1b056f-58dd-492b-8a03-6fa643cc49b8)

</details>

## Why It's Good For The Game

There aren't any ships that use a turbine yet, and the Shetland has the
space and fuel reserves for it and doesn't need the power output of a
TEG.

## Changelog

:cl:
add: Replaced the Shetland's TEG and combustion engines with a turbine
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

(cherry picked from commit 08b16477c92054db5f9604653484172d5c5119e4)