Crying-Sun-Camera-Fix (#4873)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
It turns out, the reason why you could only view 5-6 cameras on the
Crying Sun's camera network was because many of them were set to a
different network. This PR changes the incorrectly networked cameras to
run on the same network as the rest of the ship.

![Screenshot 2025-05-23
171739](https://github.com/user-attachments/assets/650b9eb4-5238-4b4c-8808-02d315705e9a)
![Screenshot 2025-05-23
171729](https://github.com/user-attachments/assets/74b3a413-3f07-4d88-a5a3-5d1793b6b2dc)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
GEC spyware on the PGF warship seems like a bad idea. I think what
happened was the original mapper might have used a camera from another
ship while mapping and forgotten to update the network tag.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: sets cameras to the correct network on the Crying Sun
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>
Co-authored-by: Changelogs <action@github.com>
Co-authored-by: Gristlebee <56049844+Gristlebee@users.noreply.github.com>
(cherry picked from commit 6b5ed5a8aab9dfa792dbdf5768c04f16261f680f)